+++ app/src/main/res/raw/changelog.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Actual fix for incorrect wind direction icon
 
+### Changed
+
+- Use DiffUtil on weather models
+
 ## [1.2.7] -- 2019-12-16
 
 ### Fixes